Move UmlDiagram.animation to parent class (TitledDiagram).
